Update 5
Cleaned up code in a few different places.

Added our to generation checker early to make it easier for later.

Changed the cell math in paint to floats.

Created the pens/fonts/brushes /stringFormats needed for all the visuals.

Added all our checks for the paint all information to the screen only if the thing is turned on.

Made sure every brush and pen are disposed of and stored in a region for neatness.